Copyright formatting?
Looks like there's something funky with the encoding of this file - the copyright is not showing up properly when being read in the addin-registry. Can you make sure the file is saved with UTF-8 encoding to properly handle the extended chars? Same for the Auto Number addin (which showed a message to that effect when editing).